refactor: rename image-gallery-queries
