+++ src/queries/image-gallery-queries.tsx
